raspberry-pi-monitoring-e-paper/requirements.txt
psutil == 6.1.0
matplotlib == 3.9.2
spidev == 3.6
gpiozero == 2.0.1
pillow == 11.0.0
numpy == 2.1.2

-e ./waveshare-source/python